// src/enums/HttpStatusCode.ts
enum HttpStatusCode {
  SuccessRequest = 200,
  CreateRequest = 201,
  NoContent = 204,
  BadRequest = 400,
  Unauthorized = 401,
  Forbidden = 403,
  NotFound = 404,
  TooManyRequests = 429,
  InternalServerError = 500,
  NotImplemented = 501,
  ServiceUnavailable = 503,
  GatewayTimeout = 504,
}

export default HttpStatusCode;